+++ ReleaseNotes.md
@@ -7,6 +7,7 @@
 * [Issue #987](https://github.com/AmpersandTarski/Ampersand/issues/987) Fixed a bug in the Excel parser.
 * refactoring Archimate parser
 * [Issue #1321](https://github.com/AmpersandTarski/Ampersand/issues/1321) Stricter testing of the prettyprinter/parser roundtrip.
+* [Issue #1090](https://github.com/AmpersandTarski/Ampersand/issues/1090) Fixed a bug in the parser for DateTime values ([UTCTime](https://hackage.haskell.org/package/time-1.13/docs/Data-Time-Format-ISO8601.html#t:ISO8601) values are supported now).
 
 ## v4.7.0 (18 June 2022)
 * [Issue #1293](https://github.com/AmpersandTarski/Ampersand/issues/1293) Fixed a bug in expressions of the form `r;V;s`